--- a.json
@@ -1,45 +1,0 @@
-{
-  "name": "Immortal",
-  "description": "nostr relay for scale",
-  "pubkey": "52b4a076bcbbbdc3a1aefa3735816cf74993b1b8db202b01c883c58be7fad8bd",
-  "software": "https://github.com/dezh-tech/immortal",
-  "supported_nips": [1, 2, 4, 9, 11, 12, 16, 20, 22, 28, 33, 40],
-  "version": "0.0.1",
-  "limitation": {
-    "max_message_length": 16384,
-    "max_subscriptions": 20,
-    "max_filters": 100,
-    "max_limit": 5000,
-    "max_subid_length": 100,
-    "max_event_tags": 100,
-    "max_content_length": 8196,
-    "min_pow_difficulty": 30,
-    "auth_required": true,
-    "payment_required": true,
-    "restricted_writes": true,
-    "created_at_lower_limit": 31536000,
-    "created_at_upper_limit": 3
-  },
-  "retention": [
-    { "kinds": [0, 1, [5, 7], [40, 49]], "time": 3600 },
-    { "kinds": [[40000, 49999]], "time": 100 },
-    { "kinds": [[30000, 39999]], "count": 1000 },
-    { "time": 3600, "count": 10000 }
-  ],
-  "tags": ["sfw-only", "bitcoin-only", "anime"],
-  "posting_policy": "https://example.com/posting-policy.html",
-  "icon": "https://nostr.build/i/53866b44135a27d624e99c6165cabd76ac8f72797209700acb189fce75021f47.jpg",
-  
-  "language_tags": ["en", "en-419"],
-  "relay_countries": ["CA", "US"],
-  "payments_url": "https://eden.nostr.land",
-  "fees": {
-    "subscription": [
-      {
-        "amount": 2500000,
-        "unit": "msats",
-        "period": 2592000
-      }
-    ]
-  }
-}
