+++ config/parameters.go
@@ -11,6 +11,21 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
+type Retention struct {
+	Time  int         `bson:"time,omitempty"  json:"time,omitempty"`
+	Count int         `bson:"count,omitempty" json:"count,omitempty"`
+	Kinds interface{} `bson:"kinds,omitempty" json:"kinds,omitempty"`
+}
+
+type Fees struct {
+	Amount int    `bson:"amount" json:"amount"`
+	Unit   string `bson:"unit"   json:"unit"`
+	Period int    `bson:"period" json:"period"`
+}
+
+type Subscription struct {
+	Subscription []Fees `bson:"subscription" json:"subscription"`
+}
 type Parameters struct {
 	Handler        *handler.Config `bson:"handler"         json:"handler"`
 	Retention      *Retention      `bson:"retention"       json:"retention"`
